% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{appendBedSuperTrack}
\alias{appendBedSuperTrack}
\title{Write a superTrack.}
\usage{
appendBedSuperTrack(track, shortLabel, longLabel, tracks, shortLabels,
  longLabels, color, priority = 2, out_file)
}
\keyword{internal}

