% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ATACseq.multiseq.stat.null}
\alias{ATACseq.multiseq.stat.null}
\title{Empirical distribution of the multiseq test statistic from ATAC-seq analysis in Shim et al, 2021.}
\format{
A vector of length 242714.
}
\description{
This is the output from our ATAC-seq analysis in Shim et al, 2021. We applied multiseq to two controls 
(media vs. ethanol) for the 242,714 regions and this data set contains the resulting 242,714 test statistics. 
The two controls are expected to have no differences in chromatin accessibility. Thus, we can construct the empirical 
null distribution of the multiseq test statistic using the 242,714 test statistics in this data set.
}
\keyword{datasets}
