% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{bamToBigWig}
\alias{bamToBigWig}
\title{Convert \code{bam} file to \code{bigWig} only extracting counts from a specified genomic region or a chromosome. By default both ends of a paired end read are counted as independent reads.}
\usage{
bamToBigWig(bam_track, chrom_file, region, bigWig_track, onlyoneend)
}
\arguments{
\item{bam_track:}{path to the \code{bam} track.}

\item{chrom_file:}{path to the tabulated file containing chromosome names and lengths}

\item{region:}{a string specifying a genomic region; extract counts from this region; can be a chromosome (e.g. "chr1") or "chr1:3456-3456789".}

\item{bigWig_track:}{name of new \code{bigWig} track}

\item{onlyoneend:}{a bool, defaults to FALSE; use TRUE if input is in bam format and only first end of the paired end read should be used.}
}
\value{
no return; it prints to a file
}
\description{
This function requires the executables \code{wigToBigWig} and \code{samtools} to be in the user's PATH. It converts \code{bam_track} into \code{bigWig_track.bw} using data from \code{chr} and \code{chromosome_file}, a tabulated file containing chromosome names and lengths.
}
\examples{
\dontrun{ 
bamToBigWig("in.bam", "chromosome_file", "chr1:3456-3456789", "out", onlyoneend=TRUE)
}
}
\keyword{internal}

