% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{bintest}
\alias{bintest}
\title{Returns estimates of intercept and slope as well as their SEs, given other input options. Called in glm.approx}
\usage{
bintest(x, g, minobs = 1, pseudocounts = 0.5, all = FALSE,
  forcebin = FALSE, repara = FALSE)
}
\arguments{
\item{x:}{a 2n by 1 vector, with first n observations giving number of successes, and next n giving number of failures in a series of binomial experiments.}

\item{g:}{covariate. Can be null, in which case only the intercept estimate and its SE is returned}

\item{minobs:}{see glm.approx}

\item{pseudocounts:}{see glm.approx}

\item{all:}{see glm.approx}

\item{forcebin:}{see glm.approx}

\item{repara:}{see glm.approx}
}
\value{
a vector of intercept and slope estimates and their SEs
}
\description{
Returns estimates of intercept and slope as well as their SEs, given other input options. Called in glm.approx
}
\keyword{internal}

