% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chr1.568739.569762.ATACseq}
\alias{chr1.568739.569762.ATACseq}
\title{ATAC-seq data over chr1:568739-569762}
\format{
A list with elements: \code{x} a 6 by 2^10 matrix of counts, \code{read.depth}
a 6-dimensional vector specifiying the total number of counts per sample, \code{g} a 
6-dimensional vector specifying the group of each sample (copper treated samples vs control samples),
\code{region} a string specifying the region of length 2^10 that reads were mapped to, \code{overall.logLR} 
a number indicating log likelihood ratio for overall expression obtained from DESeq2 output, 
\code{overall.effect} a number indicating the effect size for overall expression obtained from DESeq2 output, 
\code{overall.effect.var} a number indicating the variance of the effect size obtained from DESeq2 output.
}
\description{
This is a subset of the data collected in an ATAC-seq experiment on 3 copper treated 
samples and 3 control samples (Shim et al. 2021). Specifically, it consists of a matrix of counts 
for 6 samples over a region of length 2^10 (chr1:568739-569762) where chromatin accessibility 
is differentially expressed between two conditions (see Figure 3 in Shim et al. 2021).
}
\keyword{datasets}
