% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{chr17.10160989.10162012.DNase.seq}
\alias{chr17.10160989.10162012.DNase.seq}
\title{DNase seq data over chr17:10160989-10162012}
\format{
A list with elements: \code{x} a 70 by 2^10 matrix of counts, \code{read.depth}
a 70-dimensional vector specifiying the total number of counts per sample, a string \code{SNP}
specifying the SNP name, \code{g} a 70-dimensional vector specifying the genotype at the SNP for each sample, 
\code{region} a string specifying the region of length 2^10 that reads were mapped to.
}
\description{
This is a subset of the data collected in a DNase-seq experiment on
lymphoblastoid cell lines derived from unrelated Nigerian individuals (Degner et al., Nature
482, 390–4, 2012). Specifically, it consists of a matrix of counts for 70
samples over a region of length 2^10 (chr17:10160989-10162012) where SNP chr17.10161485 
was found to be associated with chromatin accessibility (see Figure 2 in Shim and Stephens, 
Ann. Appl. Stat. 9 665–686, 2015). DNase-seq reads and genotype data were downloaded from 
http://eqtl.uchicago.edu/dsQTL_data/.
}
\keyword{datasets}
