% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.approx.z}
\alias{compute.approx.z}
\title{Compute a vector of logit(p) given a vector of successes and failures, as well as its variance estimates}
\usage{
compute.approx.z(x.s, x.f, bound, eps, pseudocounts, all, indexn = NULL,
  return.p = FALSE)
}
\value{
a list with elements "mu", "var" and optionally "p"
}
\description{
Compute a vector of logit(p) given a vector of successes and failures, as well as its variance estimates
}

