% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{compute.dispersion}
\alias{compute.dispersion}
\title{Computes the dispersion parameter when fitting glm}
\usage{
compute.dispersion(p, n, ng, indexnm, forcebin, ind = NULL, ord = NULL,
  lg = NULL, x = NULL, x.s = NULL, x.f = NULL)
}
\value{
a vector of dispersion parameters for each fitted model, or 1 if dispersion is absent
}
\description{
Computes the dispersion parameter when fitting glm
}
\keyword{internal}

