% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.glm}
\alias{compute.glm}
\title{compute.glm}
\usage{
compute.glm(x, g, d, n, na.index, repara)
}
\value{
a matrix with estimates for mu and beta, as well as their SEs. Optionally returns "mbvar" if specified
}
\description{
compute.glm
}
\keyword{internal}

