% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compute.lm}
\alias{compute.lm}
\title{compute.lm}
\usage{
compute.lm(g, coef, se, mbvar, n, index, repara)
}
\value{
a matrix with estimates for mu and beta, as well as their SEs. Optionally returns "mbvar" if specified
}
\description{
compute.lm
}

