% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estBetaParams}
\alias{estBetaParams}
\title{Compute parameters in beta distribution.}
\usage{
estBetaParams(mu.orig, var)
}
\arguments{
\item{mu.orig}{a vector of mean for beta distribution}

\item{var}{a vector (or scalar) of variance for beta distribution}
}
\value{
a list of alpha and beta
}
\description{
This function takes mean and variance for beta distribution and returns alpha and beta parameters in beta distribution.
I copied the original version from http://stats.stackexchange.com/questions/12232/calculating-the-parameters-of-a-beta-distribution-using-the-mean-and-variance
Then I modified to handle cases when somme elements of mu are 0 or 1 (then it returns NA if mu <= 0 or >=1) or when computed alpha and beta <= 0 (then it returns NA)
}
