% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example1}
\alias{example1}
\title{A list containing counts, library size and covariates for 4 samples.}
\format{A list with elements: \code{x} a 4 by 2^14 matrix of counts (4 is the
number of samples and 2^14 is the number of adjacent bases where reads have been
counted),\code{read.depth} a 4-dimensional vector specifiying the total number
of counts per sample, and \code{g} a 4-dimensional vector specifying a
binary covariate for each sample.}
\description{
We generated this data as follows. We downloaded Encode ChipSeq
data and ran multiseq on 4 samples belonging to 2 different experiments to get
an estimated baseline and effect. Then we generated counts using a negative
binomial model using the null (the estimated baseline) and the alternative
(the estimated baseline + the estimated effect), using 2 samples for each case.
}
\keyword{datasets}

