% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example2}
\alias{example2}
\title{A list containing counts, location of the counts for 4 samples, as well
as specifications about the samples.}
\format{A list containing \code{x}, a matrix of counts for 4 samples,
\code{region} specifying the genomic location of the counts, and \code{samples}
with details about the data.}
\description{
We generated the data as follows. We downloaded Encode RNASeq
data in \code{bam} format. Then we simulated data using a betabinomial model
and thinned a region (specified in file \code{/extdata/sim/peaks.bb})
by a factor of 2 to get a differential expression pattern.
}
\keyword{#'}
\keyword{datasets}

