% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{ffwrapper}
\alias{ffwrapper}
\title{A wrapper for code in \code{compute.res} and \code{compute.res.rate}}
\usage{
ffwrapper(mean, var, Mean, Sq, wSq, beta.tm, beta.fm, g)
}
\description{
A wrapper for code in \code{compute.res} and \code{compute.res.rate}
}
\keyword{internal}

