% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.areas.strong.effects}
\alias{get.areas.strong.effects}
\title{Provide information on areas with strong effects from \code{\link{multiseq}}.}
\usage{
get.areas.strong.effects(res, threshold = 2, region = NULL)
}
\arguments{
\item{res:}{\code{\link{multiseq}} output.}

\item{threshold:}{a multiplier of the standard deviation; default is 2}

\item{region:}{a string specifying a genomic region: reference sequence name, start position, end position; defaults to NULL; if provided, the function will output the interval in genomic coordinates.}
}
\value{
This function returns a list with elements \cr
\item{chr}{a string specifying the chromosome}
\item{start}{a vector specifying the start of each area}
\item{end}{a vector specifying the end of each area}
\item{sign}{can be "+" or "-" and indicates the sign of the effect} 
\item{threshold}{multiplier of the standard deviation}
Output interval is in \code{bed} format (\code{start} is 0-based, \code{end} is 1-based).
}
\description{
This function will provide information on areas with strong effects from \code{\link{multiseq}},
 i.e., when zero is outside of the interval constructed by posterior mean +/- \code{threshold} * posterior standard 
 deviation).
}
\examples{
\dontrun{
data(chr17.10160989.10162012.DNase.seq, package="multiseq")
res <- multiseq(x=chr17.10160989.10162012.DNase.seq$x, g=chr17.10160989.10162012.DNase.seq$g)
res$region <- chr17.10160989.10162012.DNase.seq$region
plot(res, threshold = 4)
get.areas.strong.effects(res, threshold = 4)
}
}
