% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.effect.intervals}
\alias{get.effect.intervals}
\title{get.effect.intervals
intervals are printed as bed files, i.e., start is 0-based}
\usage{
get.effect.intervals(res, fra)
}
\description{
get.effect.intervals
intervals are printed as bed files, i.e., start is 0-based
}

