% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.effect.length}
\alias{get.effect.length}
\title{Return the total lenght of intervals where \code{\link{multiseq}} found strong effect (zero is outside of +/- \code{fra} * posterior standard deviation).}
\usage{
get.effect.length(res, fra)
}
\arguments{
\item{res:}{\code{\link{multiseq}} output}

\item{fra:}{a multiplier of the standard deviation; this function will return the length of intervals where \code{\link{multiseq}} found strong effect (zero is outside of +/- \code{fra} * posterior standard deviation).}
}
\description{
Return the total lenght of intervals where \code{\link{multiseq}} found strong effect (zero is outside of +/- \code{fra} * posterior standard deviation).
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr1:154206209-154214400"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

fra=2
get.effect.length(res, fra))
}
}

