% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.intervals}
\alias{get.intervals}
\title{Print intervals where \code{\link{multiseq}} found strong effect or strong peaks.}
\usage{
get.intervals(res, z.threshold = 2, p.threshold = 1e-09, region = NULL,
  what = "effect")
}
\arguments{
\item{res:}{\code{\link{multiseq}} output.}

\item{z.threshold:}{a multiplier of the standard deviation; default is 2}

\item{p.threshold:}{this argument is only used when \code{what=="baseline"} or \code{what=="log_baseline"} to specify a threshold for the detection of peaks: if \code{res$baseline.mean-z.threshold*sqrt(baseline.var)>log(p.threshold)} then a peak is called; default is 1e-09.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position, end position; defaults to NULL; if provided, the function will output the interval in genomic coordinates.}

\item{what:}{a string, it can be either "baseline" or "log_baseline" or "effect"; default is "effect"}
}
\value{
This function returns a list with elements \cr
\item{chr}{a string specifying the sequence}
\item{start}{a vector specifying the start of each interval}
\item{end}{a vector specifying the end of each interval}
\item{sign}{can be "+" or "-" and indicates the sign of the effect or is always positive when \code{what="baseline"}} 
\item{z.threshold}{multiplier of the standard deviation}
\item{p.threshold}{threshold for peak detection} 
\item{type}{where \code{type} is either "local" - if \code{res$region} or \code{region} are not specified - or "sequence" otherwise.}
Output interval is in \code{bed} format (\code{start} is 0-based, \code{end} is 1-based).
}
\description{
If \code{what=="effect"} this function will print intervals where \code{\link{multiseq}} found strong effect, i.e., when zero is outside of +/- \code{z.threshold} * posterior standard deviation). If  \code{what=="baseline"} or \code{what=="log_baseline"}  this function will print intervals where \code{\link{multiseq}} found strong peaks at a specified threshold, i.e.,  when p.threshold is below +/- \code{z.threshold} * posterior standard deviation of the log baseline.
}
\examples{
\dontrun{
data(dat, package="multiseq")
res <- multiseq(x=dat$x, g=dat$g, minobs=1, lm.approx=FALSE, read.depth=dat$read.depth)
get.intervals(res, what="effect", region=dat$region)
}
}

