% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.transcripts}
\alias{get.transcripts}
\title{Extract transcript annotation (in a specified genomic region) from a file in \code{GenePred}.
format (a standard gene annotation format).}
\usage{
get.transcripts(GenePredIn, region)
}
\arguments{
\item{GenePredIn:}{a file in \code{GenePred} format containing gene annotation}

\item{region:}{a string specifying a genomic region}
}
\description{
Extract transcript annotation (in a specified genomic region) from a file in \code{GenePred}.
format (a standard gene annotation format).
}
\examples{
GenePredIn <- file.path(path.package("multiseq"),"extdata","sim","hg19.ensGene.part.gp")
region     <- "chr1:154206209-154214400"
get.transcripts(GenePredIn, region)
}

