% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTranscripts}
\alias{getTranscripts}
\title{getTranscripts}
\usage{
getTranscripts(GenePredIn, region)
}
\arguments{
\item{GenePredIn:}{a file in GenePred format containing gene annotation}

\item{region:}{a string specifying a genomic region}
}
\description{
This function extracts transcript annotation from a file in GenePred format ithat overlap a specified region
}
\examples{
\dontrun{
GenePredIn="./knownGene_hg18.txt"
region="chr5:131989505-132120576"
getTranscripts(GenePredIn, region)
}
}

