% Generated by roxygen2 (4.0.0): do not edit by hand
\name{glm.approx}
\alias{glm.approx}
\title{This function essentially fits the model specified in documentation, using either a weighted least squares approach or a generalized linear model approach, with some modifications. It is optimized for fitting multiple models simultaneously}
\usage{
glm.approx(x, g = NULL, minobs = 1, pseudocounts = 0.5, all = FALSE,
  eps = 1e-08, center = FALSE, repara = FALSE, forcebin = FALSE,
  lm.approx = FALSE, disp = c("add", "mult"))
}
\arguments{
\item{x:}{a matrix of successes and failures for each individual at each location and each scale}

\item{g:}{a vector of covariates. Can be a factor (2 groups only) or quantitative}

\item{pseudocounts:}{adds pseudocounts in the case of low success of failure counts.}

\item{all:}{bool,  when success or failure counts are low, controls whether or not to add pseudocounts to all observations or only the low counts.}

\item{eps:}{adds tiny pseudocounts in the case of 0 successes or 0 failures. Only used when pseoducounts=0}

\item{center:}{bool, specifies if g should be centered or not}

\item{repara:}{bool, specifies if model should be reparametrized so that muhat and betahat are uncorrelated}

\item{forcebin:}{bool, if TRUE don't allow for overdipersion.}

\item{lm.approx:}{bool, specifies if WLS should be used (TRUE), or GLM (FALSE)}

\item{disp:}{"all" or "mult", indicates which type of overdispersion is assumed when lm.approx=TRUE}
}
\value{
a 4 by n matrix of muhats, SEs of muhats, betahats and SEs of betahats when repara=TRUE, and a 5 by n matrix of muhats, SEs of muhats, betahats, SEs of betahats and the covariance between muhats and betahats otherwise
}
\description{
This function essentially fits the model specified in documentation, using either a weighted least squares approach or a generalized linear model approach, with some modifications. It is optimized for fitting multiple models simultaneously
}

