% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glm.coef}
\alias{glm.coef}
\title{Compute estimates and standard errors for mu and beta when fitting WLS, as well as the covariance between mu and beta.}
\usage{
glm.coef(z, g, n, center, repara)
}
\value{
a list elements "coef", "se" and optionally "mbvar" if lg=2 and repara=TRUE, or "covv" if lg=3
}
\description{
Compute estimates and standard errors for mu and beta when fitting WLS, as well as the covariance between mu and beta.
}
\keyword{internal}

