% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{haar.aggregate}
\alias{haar.aggregate}
\title{Get data at all scales.}
\usage{
haar.aggregate(x)
}
\arguments{
\item{x:}{nind by 2^k matrix of Poisson counts}
}
\value{
a nind by (2*2^k)-2 matrix that sums up columns of x at different scales
}
\description{
This function takes a matrix x of Poisson counts, nind by 2^k, and returns a nind by (2*2^k)-2 matrix that sums up columns of x at different scales. The last two columns of the aggregate are the sums of the first half of x and the sums of the second half. This function is needed to estimate the pi-s by aggregating data from multiple regions.
}
\keyword{internal}

