% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{intervalsToBigBed}
\alias{intervalsToBigBed}
\title{Write a \code{bigBed} file given the output of \code{get.intervals}}
\usage{
intervalsToBigBed(intervals, chrom_file, bigBedFile = "out.bb")
}
\arguments{
\item{intervals:}{output of get.intervals}

\item{chrom_file:}{path to the file containing chromosome names and lengths}

\item{bigBedFile:}{output file, defaults to "out.bb"}
}
\keyword{internal}

