% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.multiseq}
\alias{plot.multiseq}
\title{Plot the output of \code{\link{multiseq}} (either the effect or the baseline).}
\usage{
\method{plot}{multiseq}(
  x,
  is.xaxis = TRUE,
  threshold = 2,
  what = "effect",
  highlight = TRUE,
  col.mean = "navy",
  col.sd = "sky blue",
  axes = F,
  type = "l",
  main = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  cex = NULL,
  ...
)
}
\arguments{
\item{col.mean}{: a string, specifying the color for posterior mean.}

\item{col.sd}{: a string, specifying the color for +/- \code{threshold} * posterior standard deviations.}

\item{x:}{multiseq output; if x$region is defined then the \code{x} axis will use genomic coordinates.}

\item{is.xaxis:}{bool, if TRUE plot \code{x} axis otherwise don't plot \code{x} axis.}

\item{threshold:}{a multiplier of the standard deviation.}

\item{what:}{a string, it can be either "baseline" or "log_baseline" or "effect".}

\item{highlight:}{a bool, if TRUE, highlight areas with strong signal when what == "effect"; defaults to TRUE;}
}
\description{
If `what=="effect"` this function will plot the posterior mean (line in \code{col.mean}) for effect size,
 i.e., the difference in log-intensity between groups, and +/- \code{threshold} * posterior standard deviations (lines in \code{col.sd}). 
 If `what=="effect"` and `highlight==TRUE`, this function will highlight areas with strong effects (i.e., zero is 
 outside of the interval constructed by the two lines in \code{col.sd}). If  `what=="baseline”`, this function will plot the posterior mean 
 for the baseline \code{exp(res$baseline.mean)}. If `what=="log_baseline”`, this function will plot the posterior mean 
 for the log of the baseline \code{res$baseline.mean}, and +/- \code{threshold} * posterior standard deviations. 
 If x$region is defined then the \code{x} axis will use genomic coordinates.
}
\examples{
\dontrun{
data(chr17.10160989.10162012.DNase.seq, package="multiseq")
res <- multiseq(x=chr17.10160989.10162012.DNase.seq$x, g=chr17.10160989.10162012.DNase.seq$g)
res$region <- chr17.10160989.10162012.DNase.seq$region
plot(res, threshold = 3)
plot(res, what="baseline")
}
}
