% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.transcripts}
\alias{plot.transcripts}
\title{Plot transcripts; if \code{region} is specified then plot only the specified region.}
\usage{
\method{plot}{transcripts}(Transcripts, region = NULL, is.xaxis = TRUE,
  axes = F, xlim = NULL, type = NULL, cex = NULL, font.main = 1, ...)
}
\arguments{
\item{Transcripts:}{output of \code{\link{get.transcripts}}}

\item{region:}{region to be plotted; defaults to NULL.}

\item{is.xaxis:}{bool, if TRUE plot \code{x} axis otherwise don't plot \code{x} axis; defaults to TRUE.}
}
\description{
The only required field for this function is Transcripts.
}
\examples{
region     <- "chr12:113354417-113358512" 
GenePredIn  <- file.path(path.package("multiseq"),"extdata","hg19.OAS1.refGene.part.gp") 
Transcripts <- get.transcripts(GenePredIn, region)
plot(Transcripts, region)
}

