% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotResults}
\alias{plotResults}
\title{plotResults}
\usage{
plotResults(res, fra = 2, type = "effect", title = NULL, ylim = NULL)
}
\arguments{
\item{res:}{multiseq output}

\item{fra:}{a multiplier of the standard deviation}

\item{type:}{if type is "baseline" plot multiseq baseline output; if type is "effect" plot multiseq efefct output; defaults to "effect"}

\item{title:}{the title of the plot; defaults to NULL}

\item{ylim:}{plot ylim; defaults to default ylim}
}
\description{
This function plots the output of multiseq (either the effect or the baseline) and *fra* * posterior standard deviation
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr5:131989505-132120576"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)
plotResults(res)
plotResults(res,"baseline")
}
}

