% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotTranscripts}
\alias{plotTranscripts}
\title{plotTranscripts}
\usage{
plotTranscripts(Transcripts, plotStart = NULL, plotEnd = NULL,
  is.xaxis = 1, main = NULL, cex = 1)
}
\arguments{
\item{Transcripts:}{output of getTranscripts}

\item{plotStart:}{minimum value for x axis}

\item{plotEnd:}{maximum value for x axis}

\item{is.xaxis:}{bool, if TRUE plot x axis otherwise don't plot x axis}

\item{main:}{string, title}

\item{cex:}{number indicating the amount by which plotting text and symbols should be scaled relative to the default. 1=default, 1.5 is 50% larger, 0.5 is 50% smaller, etc.}
}
\description{
plotTranscripts
}
\examples{
\dontrun{
GenePredIn  <- "./knownGene_hg18.txt"
region      <- "chr5:131989505-132120576"
Transcripts <- getTranscripts(GenePredIn, region)
plotTranscripts(Transcripts)
}
}

