% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{printGoToMessage}
\alias{printGoToMessage}
\title{Write a message containing instructions on how to visualize the Track Hub.}
\usage{
printGoToMessage(hub_name, assembly, hub_dir, http_address, region = NULL)
}
\keyword{internal}

