% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reflect}
\alias{reflect}
\title{Reflects the signal and extends it to a power of 2 if necessary}
\usage{
reflect(x)
}
\arguments{
\item{x:}{an m by n matrix of signals}
}
\value{
an n-vector containing the indices of the smoothed output corresponding to the original signal x. Also substitutes the original signal with the reflected one to be used as input.
}
\description{
Reflecting the signal in wavelet applications allows the signal to be periodic and usually lessens boundary bias. If the original signal
has a length which is a power of 2, this function adds a reflection of the original signal to itself. Otherwise this function extends
the original signal to have length of a power of 2 by reflecting an appropriate number of observation points, and then reflecting the entire modified signal.
}
\keyword{internal}

