% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{reverse.pwave}
\alias{reverse.pwave}
\title{Reverse wavelet transform a set of probabilities in TItable format}
\usage{
reverse.pwave(est, lp, lq = NULL)
}
\arguments{
\item{est:}{an n-vector. Usually a constant vector with each element equal to the estimated log(total rate).}

\item{lp:}{a J by n matrix of estimated log(p), wavelet proportions for Poisson data on the log scale. First row of lp gives the high frequency proportions (so 1/(1+2), 3/(3+4) etc), second row gives next level of resolution etc.}

\item{lq:}{a J by n matrix of estimated log(1-p). If lq is not given, set lq = log(1-exp(lp)).}
}
\value{
an n-vector
}
\description{
Reverse wavelet transform a set of probabilities in TItable format
}
\keyword{internal}

