% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{safe.quasibinomial.glm.fit}
\alias{safe.quasibinomial.glm.fit}
\title{Modified glm function to return relevant outputs, not allowing for underdispersion}
\usage{
safe.quasibinomial.glm.fit(x, y, forcebin = FALSE, repara = FALSE, ...)
}
\arguments{
\item{x:}{covariante}

\item{y:}{response}

\item{forcebin:}{see glm.approx}

\item{repara:}{see glm.approx}

\item{...:}{other inputs to glm.fit}
}
\value{
a vector of intercept and slope estimates and their SEs
}
\description{
Modified glm function to return relevant outputs, not allowing for underdispersion
}
\keyword{internal}
