% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample.from.Binomial.with.Overdispersion}
\alias{sample.from.Binomial.with.Overdispersion}
\title{Simulates binomial samples with/without over dispersion.}
\usage{
sample.from.Binomial.with.Overdispersion(
  num.sam,
  total.count,
  mu.sig,
  over.dispersion = NULL
)
}
\arguments{
\item{num.sam}{number of samples to be sampled}

\item{total.count}{a vector of non-negative counts;}

\item{mu.sig}{a vector of probabilities (we allow 0 or 1 as probablity)}

\item{over.dispersion}{if over.dispersion == NULL, simulate data from binomial. If over.dispersion is provided, simulate data binomial with over.dispersion.}
}
\value{
a matrix of num.sam by L (length of total.count) containing simulated data.
}
\description{
This function simulate samples from binomial distribution or beta-binomial distribution 
as described in Shim et al., 2021.
}
