% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{setAshParam}
\alias{setAshParam}
\title{Set default \code{ash} parameters.}
\usage{
setAshParam(ashparam)
}
\arguments{
\item{ashparam:}{a list of parameters to be passed to ash.}
}
\description{
Set default \code{ash} parameters.
}
\keyword{internal}

