% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{setLearnPiParam}
\alias{setLearnPiParam}
\title{Set default \code{learn.pi} parameters.}
\usage{
setLearnPiParam(learn.pi.param)
}
\arguments{
\item{learn.pi.param:}{a list of parameters to be passed to learn.pi.}
}
\description{
Set default \code{learn.pi} parameters.
}
\keyword{internal}

