% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{suppressW}
\alias{suppressW}
\title{Suppress an ash warning that we want to ignore; used any time ash is called in the package}
\usage{
suppressW(w)
}
\description{
Suppress an ash warning that we want to ignore; used any time ash is called in the package
}
\keyword{internal}

