% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{tfmoment}
\alias{tfmoment}
\title{Compute the third and fourth moments of a normal or a mixture of normals distribution.}
\usage{
tfmoment(mu, sigma, moment, pi = NULL)
}
\arguments{
\item{moment:}{if moment=3 compute the third moment, if moment=4 compute the forth moment}
}
\description{
Compute the third and fourth moments of a normal or a mixture of normals distribution.
}
\keyword{internal}

