% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{vectorToBigWig}
\alias{vectorToBigWig}
\title{Write a vector of counts to a \code{bigWig} file given position and sequence information}
\usage{
vectorToBigWig(x, chr, start, chrom_file, bigWigFile = "out.wg")
}
\arguments{
\item{x:}{an R vector of (per base) counts}

\item{chr:}{sequence name}

\item{start:}{position on the sequence "chr" where the counting process started}

\item{chrom_file:}{path to the file containing chromosome names and lengths}

\item{bigWigFile:}{path to the output \code{bigWig} file, defaults to "out.wg"}
}
\keyword{internal}

