% Generated by roxygen2 (4.0.0): do not edit by hand
\name{wls.coef}
\alias{wls.coef}
\title{Compute estimates and standard errors for mu and beta when fitting WLS, as well as the covariance between mu and beta}
\usage{
wls.coef(z, disp, indexnm, n, ng, forcebin, g = NULL, repara = NULL)
}
\value{
a list elements "coef", "se", "mbvar"
}
\description{
Compute estimates and standard errors for mu and beta when fitting WLS, as well as the covariance between mu and beta
}

