% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wls.mb}
\alias{wls.mb}
\title{wls.mb}
\usage{
wls.mb(z, v, disp, indexnm, ng, df, forcebin, g = NULL, n = NULL,
  vv = NULL)
}
\value{
a list with elements "coef", "se", "wrse2" if g is not specified, or a list with elements "muhat", "semuhat", "betahat", "sebetahat", "covmubeta", "wrse2" otherwise
}
\description{
wls.mb
}

