% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.bed}
\alias{write.bed}
\title{Write effect intervals to a \code{bed} file.}
\usage{
write.bed(intervals, bedfile)
}
\arguments{
\item{intervals:}{output of \code{\link{get.effect.intervals}}}

\item{bedfile:}{output \code{bed} file}
}
\description{
Write effect intervals to a \code{bed} file.
}
\examples{
\dontrun{
fra=2
region="chr1:154206209-154214400"
res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)
intervals <- get.effect.intervals(res, fra))
write.bed(intervals, "out.bed")
}
}

