% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.effect.intervals}
\alias{write.effect.intervals}
\title{write.effect.intervals}
\usage{
write.effect.intervals(res, dir.name, fra = 2, region = NULL)
}
\arguments{
\item{intervals:}{output of get.effect.intervals}

\item{dir.name:}{output directory}

\item{fra:}{a multiplier of the standard deviation; this function will write intervals where there is an effect at plus or minus fra * standard deviation.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position (locus.start), end position (locus.end); defaults to NULL}
}
\description{
Write intervals with strong effect (zero is outside of +/- *fra* * posterior standard deviation) to bed file
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr5:131989505-132120576"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

write.effect.intervals(res, "~/output_folder", fra=2, region))
}
}

