% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.effect.mean.variance.gz}
\alias{write.effect.mean.variance.gz}
\title{write.effect.mean.variance.gz}
\usage{
write.effect.mean.variance.gz(res, dir.name, type = "effect")
}
\arguments{
\item{res:}{multiseq output}

\item{fra:}{a multiplier of the standard deviation; this function will return the length of intervals where there is an effect at plus or minus fra * standard deviation.}

\item{type:}{if type is "baseline" plot multiseq baseline output; if type is "effect" plot multiseq efefct output; defaults to "effect"}
}
\description{
This function saves multiseq results in file effect_mean_var.txt.gz, a file with two columns: first column is effect (baseline) mean and second column is effect (baseline) variance
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr5:131989505-132120576"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

write.effect.mean.variance.gz(res, "~/output_folder")
}
}

