% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.gz}
\alias{write.gz}
\title{Write \code{\link{multiseq}} results to a compressed file.}
\usage{
write.gz(res, file = "results.mean.sd2.gz", what = "effect")
}
\arguments{
\item{res:}{\code{\link{multiseq}} output.}

\item{file:}{path to the output file.}

\item{what:}{if \code{what} is "baseline" then plot \code{\link{multiseq}} baseline output; if \code{what} is "effect" then plot \code{\link{multiseq}} effect output; defaults to "effect".}
}
\description{
The output file has two columns: first column the effect (or the baseline) mean and second column is the effect (or the baseline) posterior standard deviation ^2.
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr1:154206209-154214400"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Type)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

write.gz(res, "./results.mean.sd2.gz")
}
}

