% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{writeTrackHubSkeleton}
\alias{writeTrackHubSkeleton}
\title{Write \code{genomes.txt} and \code{hub.txt} files}
\usage{
writeTrackHubSkeleton(hub_dir, assembly = "hg19", hub_name_string,
  email = "fake_email_address@dot.com")
}
\keyword{internal}

