% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{appendBedSuperTrack}
\alias{appendBedSuperTrack}
\title{Write a superTrack with BigBed files in the Track Hub trackdb_file}
\usage{
appendBedSuperTrack(track, shortLabel, longLabel, tracks, shortLabels,
  longLabels, color, priority = 2, out_file)
}
\description{
Write a superTrack with BigBed files in the Track Hub trackdb_file
}

