% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bamToBigWig}
\alias{bamToBigWig}
\title{Convert \code{bam} file to \code{bigWig} only extracting counts from a specified chromosome. By default both ends of a paired end read are counted as independent reads.}
\usage{
bamToBigWig(bam_track, chrom_file, chr, track_name, onlyoneend)
}
\arguments{
\item{bam_track:}{path to the \code{bam} track.}

\item{chrom_file:}{path to the tabulated file containing chromosome names and lengths}

\item{chr:}{extract counts from the specified chromosome}

\item{bigWig_track:}{name of new \code{bigWig} track}

\item{onlyoneend:}{a bool, defaults to FALSE; use TRUE if input is in bam format and only first end of the paired end read should be used.}
}
\value{
no return; it prints to a file
}
\description{
This function requires the executables \code{wigToBigWig} and \code{samtools} to be in the user's PATH. It converts \code{bam_track} into \code{bigWig_track.bw} using data from \code{chr} and \code{chromosome_file}, a tabulated file containing chromosome names and lengths.
}
\examples{
\dontrun{
bamToBigWig("in.bam", "chromosome_file", "chr1", "out", onlyoneend=TRUE)
}
}
\keyword{internal}

