% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{compute.approx.z}
\alias{compute.approx.z}
\title{compute.approx.z}
\usage{
compute.approx.z(x.s, x.f, bound, eps, pseudocounts, all, indexn = NULL,
  return.p = FALSE)
}
\value{
a list with elements "mu", "var" and optionally "p"
}
\keyword{(MLE}
\keyword{Berkso's}
\keyword{Compute}
\keyword{Tukey's}
\keyword{a}
\keyword{and}
\keyword{approximation}
\keyword{as}
\keyword{at}
\keyword{endpoints}
\keyword{estimates}
\keyword{estimator}
\keyword{failures,}
\keyword{for}
\keyword{given}
\keyword{internal}
\keyword{its}
\keyword{logit(p)}
\keyword{mean;}
\keyword{mix}
\keyword{of}
\keyword{successes}
\keyword{variance}
\keyword{variance)}
\keyword{vector}
\keyword{well}
\keyword{with}

