% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{compute.glm}
\alias{compute.glm}
\title{compute.glm}
\usage{
compute.glm(x, g, d, n, na.index, repara)
}
\value{
a matrix with estimates for mu and beta, as well as their SEs. Optionally returns "mbvar" if specified
}
\keyword{internal}

