% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{compute.lm}
\alias{compute.lm}
\title{compute.lm}
\usage{
compute.lm(g, coef, se, mbvar, n, index, repara)
}
\value{
a matrix with estimates for mu and beta, as well as their SEs. Optionally returns "mbvar" if specified
}
\keyword{internal}

