% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseq.R
\name{compute.res.rate}
\alias{compute.res.rate}
\title{compute.res.rate}
\usage{
compute.res.rate(zdat, repara, baseline, w, read.depth, g = NULL)
}
\value{
a list with elements "lp.mean", "lp.var", "lpratio.mean", "lpratio.var"
}
\keyword{internal}

