% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compute.res.rate}
\alias{compute.res.rate}
\title{compute.res.rate}
\usage{
compute.res.rate(zdat, repara, w, read.depth, g = NULL)
}
\value{
a list with elements "lp.mean", "lp.var", "lpratio.mean", "lpratio.var"
}
\description{
compute.res.rate
}

