% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.counts}
\alias{get.counts}
\title{Prepare input for multiseq function extracting counts in a genomic region from \code{bam}, \code{hdf5}, or \code{bigWig} files.}
\usage{
get.counts(samplesheet = NULL, region = NULL, onlyoneend = FALSE)
}
\arguments{
\item{samplesheet:}{a string specifying the path to a samplesheet; the samplesheet should contain a column with header \code{SampleID} and a column with header either \code{bamPath} or \code{hdf5Path} or \code{bigWigPath}.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position, end position (see example below).}

\item{onlyonend:}{a bool, defaults to FALSE; use TRUE if input is in bam format and only first end of the paired end read should be used.}
}
\value{
a matrix with \code{N} rows and \code{end-start+1} columns containing the number of reads that start at each base in the specified region in each sample. Rownames correspond to \code{samples$SampleID}
}
\description{
This functions reads in a `samplesheet` and extracts read counts in a genomic region specified by parameter `region`, preparing input for \code{\link{multiseq}}. If in file `samplesheet` \code{samples$bamPath} is specified then this function extracts reads from the bam files in \code{samples$bamPath} using \code{samtools} (which must be in the USER's path) (no filter applied but option \code{onlyonend} can be used to extract reads from only one end if reads are paired end). Else if \code{samples$hdf5Path} is specified this function extracts reads from the \code{hdf5} files in \code{samples$hdf5Path} using the R package \code{rhdf5}. If \code{samples$bigWigPath} is specified this function extracts reads from \code{bigWig} files using the executable \code{bigWigToWig} (which must be in the USER's path).
}
\examples{
\dontrun{
setwd(file.path(path.package("multiseq"),"extdata"))
samplesheet="samplesheetEncode.txt"
region="chr1:11740409-11756792"
x=get.counts(samplesheet, region)
}
}

