% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.counts}
\alias{get.counts}
\title{Prepare input for multiseq function extracting counts in a genomic region from bam, hdf5, or bigWig files}
\usage{
get.counts(samplesheet, region, onlyoneend = FALSE)
}
\arguments{
\item{samplesheet:}{a string or a data frame of size N equal to the number of samples. If a string it should be the path to a samplesheet; the samplesheet should contain a column with header SampleID and a column with header either bigWigPath or h5FullPath or bigWigPath. If a data frame, samples$SampleID must be specified and either samples$bigWigPath or samples$h5FullPath or samples$bamReads must be specified.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position (locus.start), end position (locus.end)}

\item{onlyonend:}{a bool, defaults to FALSE; use TRUE if input is in bam format and only first end of the paired end read should be used.}
}
\value{
a matrix with N rows and locus.end-locus.start+1 columns containing the number of reads that start at each base in the specified region in each sample. Rownames correspond to samples$SampleID
}
\description{
This functions extracts read counts from *bam*, *hdf5*, or *bigWig* files in a genomic region, preparing input for multiseq. If samples$bamReads is specified then this function extracts reads from the bam files in samples$bamReads using `samtools` (which must be in the USER's path) (no filter applied). Else if samples$h5FullPath is specified this function extracts reads from the hdf5 files in samples$h5FullPath using the R package rhdf5. If samples$bigWigPath is specified this function extracts reads from bigWig files using the executable `bigWigToWig` (which must be in the USER's path).
}
\examples{
\dontrun{
setwd(file.path(path.package("multiseq"),"extdata","sim"))
samplesheet="samplesheet.sim.txt"
region="chr1:87297710-87305901"
M=get.counts(samplesheet, region)
}
}

