% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.effect.intervals}
\alias{get.effect.intervals}
\title{get.effect.intervals}
\usage{
get.effect.intervals(res, fra, region = NULL)
}
\arguments{
\item{res:}{multiseq output}

\item{fra:}{a multiplier of the standard deviation; this function will output intervals where there is an effect at plus or minus fra * standard deviation.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position (locus.start), end position (locus.end); defaults to NULL if provided, the function will output the interval in genomic coordinates}
}
\value{
a list with elements chr, start, end, sign (of the effect), fra, type (type can be "local" or "sequence" and specifies whether start and end are relative to the genomic sequence
}
\description{
This function outputs intervals where multiseq found strong effect (zero is outside of +/- *fra* * posterior standard deviation). Output interval is in bed format (start is 0-based, end is 1-based)
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr5:131989505-132120576"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

get.effect.intervals(res, fra, region))
}
}

