% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.effect.length}
\alias{get.effect.length}
\title{get.effect.length}
\usage{
get.effect.length(res, fra)
}
\arguments{
\item{res:}{multiseq output}

\item{fra:}{a multiplier of the standard deviation; this function will return the length of intervals where there is an effect at plus or minus fra * standard deviation.}
}
\description{
This function returns the total lenght of intervals where multiseq found an effect at fra * posterior standard deviation
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr5:131989505-132120576"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)

get.effect.length(res, fra=2))
}
}

