% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.transcripts}
\alias{get.transcripts}
\title{Extract transcript annotation (in a specified genomic region) from a file in \code{GenePred}.
format (a standard gene annotation format).}
\usage{
get.transcripts(GenePredIn, region)
}
\arguments{
\item{GenePredIn:}{a file in \code{GenePred} format containing gene annotation}

\item{region:}{a string specifying a genomic region}
}
\examples{
GenePredIn <- file.path(path.package("multiseq"),"extdata","hg19.OAS1.refGene.part.gp")
region     <- "chr12:113354417-113358512"
get.transcripts(GenePredIn, region)
}

