% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{glm.approx}
\alias{glm.approx}
\title{Fit the model specified in documentation, using either a weighted least squares approach or a generalized linear model approach, with some modifications.}
\usage{
glm.approx(x, g = NULL, minobs = 1, pseudocounts = 0.5, all = FALSE,
  eps = 1e-08, center = FALSE, repara = FALSE, forcebin = FALSE,
  lm.approx = FALSE, disp = c("add", "mult"), bound = 0.02)
}
\arguments{
\item{x:}{a matrix of N (# of samples) by 2*T (T: # of WCs or, more precisely, of different scales and locations in multi-scale space); Two consecutive columns correspond to a particular scale and location; The first column (the second column) contains # of successes (# of failures) for each sample at corresponding scale and location.}

\item{g:}{a vector of covariates. Can be a factor (2 groups only) or quantitative}

\item{pseudocounts:}{adds pseudocounts in the case of low success of failure counts.}

\item{all:}{bool,  when success or failure counts are low, controls whether or not to add pseudocounts to all observations or only the low counts.}

\item{eps:}{adds tiny pseudocounts in the case of 0 successes or 0 failures. Only used when pseoducounts=0}

\item{center:}{bool, specifies if g should be centered or not}

\item{repara:}{bool, specifies if model should be reparametrized so that muhat and betahat are uncorrelated}

\item{forcebin:}{bool, if TRUE don't allow for overdipersion.}

\item{lm.approx:}{bool, specifies if WLS should be used (TRUE), or GLM (FALSE)}

\item{disp:}{"all" or "mult", indicates which type of overdispersion is assumed when lm.approx=TRUE}

\item{bound:}{numeric, indicates the threshold of the success vs failure ratio below which pseudocounts will be added}
}
\value{
a matrix of 2 (or 5 if g is provided) by T (# of WCs); Each row contains alphahat (1st row), standard error of alphahat (2nd), betahat (3rd), standard error of betahat (4th), covariance between alphahat and betahat (5th) for each WC.
}
\description{
This function is optimized for fitting multiple models simultaneously
}
\keyword{internal}

