% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdf5ToBigWig}
\alias{hdf5ToBigWig}
\title{convert hdf5 file (with data from a locus) to bigwig file}
\usage{
hdf5ToBigWig(h5_track, chrom_file, bigWig_track, assembly = "hg19")
}
\arguments{
\item{h5_track:}{path to the hdf5 track}

\item{chrom_file:}{path to the file containing chromosome names and lengths}

\item{bigWig_track:}{name of new bigWig track}

\item{assembly:}{genome assembly that reads were mapped to; default="hg19"}
}
\value{
no return; it prints a Track Hub folder
}
\description{
This function requires the executable "wigToBigWig" to be in the user's PATH
}
\examples{
hdf5ToBigWig(in.h5, chromosome_file, out)
converts in.h5 into out.bw using chromosome names and lengths in chromosome_file
}

