% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intervalsToBigBed}
\alias{intervalsToBigBed}
\title{Write a \code{bigBed} file given the output of \code{get.ffect.intervals}}
\usage{
intervalsToBigBed(intervals, chrom_file, bigBedFile = "out.bb")
}
\arguments{
\item{intervals:}{output of get.effect.intervals}

\item{chrom_file:}{path to the file containing chromosome names and lengths}

\item{bigBedFile:}{output file, defaults to "out.bb"}
}
\description{
Write a \code{bigBed} file given the output of \code{get.ffect.intervals}
}
\keyword{internal}

