% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{multiseqToTrackHub}
\alias{multiseqToTrackHub}
\title{Create a UCSC genome browser "Track Hub" from \code{\link{multiseq}} output.}
\usage{
multiseqToTrackHub(res, z.threshold = NULL, hub_name = "multiseq",
  chrom_file = system.file("extdata", "chromosome.lengths.hg19.txt", package =
  "multiseq"), assembly = "hg19",
  mountpoint = Sys.getenv("MOUNTPOINT_PATH"),
  http_address = Sys.getenv("MOUNTPOINT_HTTP_ADDRESS"))
}
\arguments{
\item{res:}{output of \code{\link{multiseq}}; \code{res$region} should be
defined (e.g.: \code{res$region="chr1:2345-234567")}; a valid region must
contain sequence_name:locus_start-locus_end}

\item{z.threshold:}{a multiplier of the standard deviation; this function will create
a Truck Hub with the effect at plus or minus \code{z.threshold} * posterior standard
deviation; by default the function will use \code{z.threshold=res$z.threshold}}

\item{hub_name:}{name of the Track Hub; this string can be set to any value;
it could contain a path, in which case the path will be relative to the
mountpoint; default="multiseq"}

\item{chrom_file:}{path to the file containing chromosome names and lengths;
default is \code{system.file("extdata", "chromosome.lengths.hg19.txt", package="multiseq")}}

\item{assembly:}{genome assembly that reads were mapped to; default="hg19"}

\item{mountpoint:}{path to the directory where the Track Hub folder will be
saved in. This directory should be associated with an http address or an
ftp address; default=\code{Sys.getenv("MOUNTPOINT_PATH")}}

\item{http_address:}{http or ftp address associated with the mountpoint;
default=\code{Sys.getenv("MOUNTPOINT_PATH")}}
}
\description{
This function requires the executables \code{wigToBigWig} and
\code{bedToBigBed} to be in the user's PATH. By default this function uses
the human genome \code{hg19}. To use this function with a different genome
specify \code{chrom_file} and \code{assembly}.
}
\examples{
\dontrun{
region="chr1:154206209-154214400"
#res is the putput of multiseq
res$region <- region
multiseqToTrackHub(res)
} 
}

