% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{data}
\name{OAS1}
\alias{OAS1}
\title{A data frame containin counts, library size and covariates for 66 RNA samples with differential expression}
\format{A data frame with elements: "M" a 66 by 4096 matrix of counts at 4096 adjacent positions for 66 RNA samples, "read.depth" ang "g" vectors of length 66 specifying the library size and the genotype, respectively, for each sample.}
\description{
A data frame containin counts, library size and covariates for 66 RNA samples with differential expression
}
\keyword{datasets}

