% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.multiseq}
\alias{plot.multiseq}
\title{Plot the output of \code{\link{multiseq}} (either the effect or the baseline).}
\usage{
\method{plot}{multiseq}(x, is.xaxis = TRUE, z.threshold = 2,
  p.threshold = 1e-09, what = "effect", highlight = TRUE, axes = F,
  type = "l", col = "green", main = NULL, ylim = NULL, xlab = NULL,
  ylab = NULL, cex = NULL, ...)
}
\arguments{
\item{x:}{multiseq output; if x$region is defined then the \code{x} axis will use genomic coordinates.}

\item{is.xaxis:}{bool, if TRUE plot \code{x} axis otherwise don't plot \code{x} axis.}

\item{z.threshold:}{a multiplier of the standard deviation.}

\item{p.threshold:}{this argument is only used when \code{what=="baseline"} or \code{what=="log_baseline"} to specify a threshold for the detection of peaks: if \code{res$baseline.mean-z.threshold*sqrt(baseline.var)>log(p.threshold)} then a peak is called; defaults to 1e-09.}

\item{what:}{a string, it can be either "baseline" or "log_baseline" or "effect".}

\item{highlight:}{a bool, if TRUE highlight intervals with strong peaks or strong signal; defaults to TRUE.shell}
}
\description{
If `what=="effect"` this function will plot the effect and intervals where \code{\link{multiseq}} found strong effect, i.e., when zero is outside of +/- \code{z.threshold} * posterior standard deviation). If  `what=="baseline"` or `what=="log_baseline"` this function will plot either the baseline \code{exp(res$baseline.mean)} or the log of the baseline \code{res$baseline.mean}, respectively, and  intervals where \code{\link{multiseq}} found strong peaks at a specified threshold, i.e., when log(p.threshold) is below +/- \code{z.threshold} * posterior standard deviation of the log baseline. If x$region is defined then the \code{x} axis will use genomic coordinates.
}
\examples{
\dontrun{
data(dat, package="multiseq")
res <- multiseq(x=dat$x, g=dat$g, minobs=1, lm.approx=FALSE, read.depth=dat$read.depth)
res$region <- dat$region
plot(res)
plot(res, what="baseline")
}
}

