% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.multiseq}
\alias{plot.multiseq}
\title{Plot the output of \code{\link{multiseq}} (either the effect or the baseline) and \code{fra} * posterior standard deviation.}
\usage{
\method{plot}{multiseq}(x, region = NULL, is.xaxis = TRUE, fra = 2,
  what = "effect", axes = F, type = "l", col = "green", main = NULL,
  ylim = NULL, xlab = NULL, ylab = NULL, cex = NULL, ...)
}
\arguments{
\item{x:}{multiseq output.}

\item{region:}{a string specifying a genomic region: reference sequence name, start position, end position; defaults to NULL; if provided (or if not provided but x$region is defined) the x axis label and tickers will contain genomic information.}

\item{is.xaxis:}{bool, if TRUE plot \code{x} axis otherwise don't plot \code{x} axis.}

\item{fra:}{a multiplier of the standard deviation.}

\item{what:}{a string, it can be either "baseline" or "effect"; if "baseline", this function plots multiseq baseline output; if "baseline" this function plots multiseq effect output; defaults to "effect".}
}
\description{
Plot the output of \code{\link{multiseq}} (either the effect or the baseline) and \code{fra} * posterior standard deviation.
}
\examples{
\dontrun{
#run multiseq on sample data
samplesheet <- file.path(path.package("multiseq"), "extdata", "sim", "samplesheet.sim.txt")
region      <- "chr1:154206209-154214400"
x           <- get.counts(samplesheet, region)
samples     <- read.table(samplesheet, stringsAsFactors=F, header=T)
g           <- factor(samples$Tissue)
g           <- match(g, levels(g))-1

res <- multiseq(x=x, g=g, minobs=1, lm.approx=FALSE, read.depth=samples$ReadDepth)
plot.multiseq(res)
plot.multiseq(res, what="baseline")
}
}

