% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackHub.R
\name{samplesheetToTrackHub}
\alias{samplesheetToTrackHub}
\title{Create a UCSC Genome Browser "Track Hub" from read tracks and bed tracks (of significant intervals) listed in a \code{samplesheet}.}
\usage{
samplesheetToTrackHub(samplesheet, hub_name = paste0(basename(samplesheet),
  ".TrackHub"), chrom_file = system.file("extdata",
  "chromosome.lengths.hg19.txt", package = "multiseq"), region = NULL,
  assembly = "hg19", mountpoint = Sys.getenv("MOUNTPOINT_PATH"),
  http_address = Sys.getenv("MOUNTPOINT_HTTP_ADDRESS"), onlyoneend = FALSE)
}
\arguments{
\item{samplesheet:}{a string specifying the path to the samplesheet;
for specifications of the samplesheet format see the vignette.
A column with header \code{sampleID} is required; either a column
with header \code{hdf5Path}, or a column with header \code{bigWigPath},
or a column with header \code{bamPath} containing the path to the \code{hdf5}
files or the \code{bigWig} files or the \code{bam} files, respectively,
should be specified. If the samplesheet has a column with header \code{bedPath}
(specifying the path to a \code{bed} or \code{bigBed} file). Depending on the size of the files this code might require a lot of memory.}

\item{hub_name:}{name of the Track Hub; this string can be set to any value; it could contain a path, in which case the path will be relative to the mountpoint (see below); defaults to \code{paste0(basename(samplesheet),".TrackHub")}.}

\item{chrom_file:}{path to the file containing chromosome names and lengths; defaults \code{system.file("extdata", "chromosome.lengths.hg19.txt", package="multiseq")}.}

\item{region:}{a string, restrict ouput to the selected genomic region; can be a chromosome or a region (e.g., "chr1:1245-23456789").}

\item{assembly:}{genome assembly that reads were mapped to; defaults to "hg19".}

\item{mountpoint:}{path to the directory where the track hub folder will be saved in; this directory should be associated with an \code{http address} or an \code{ftp address}; defaults to \code{Sys.getenv("MOUNTPOINT_PATH")}.}

\item{http_address:}{http or ftp address associated with the mountpoint; defaults to \code{Sys.getenv("MOUNTPOINT_PATH")}.}

\item{onlyoneend:}{a bool, defaults to FALSE; use TRUE if input is in \code{bam} format and only first end of the paired end read should be used.}
}
\description{
This function requires the executables \code{wigToBigWig}, \code{bedToBigBed}, and \code{bigWigInfo} to be in the user's PATH.
Read tracks can be in \code{bam}, \code{hdf5}, \code{bigWig}, \code{bed} or \code{bigBed} format and significant intervals can only be in bed format. If in \code{bam} format by default both ends of a paired end read are counted as independent reads (control this behavior with parameter onlyoneend).
}
\examples{
\dontrun{
setwd(file.path(path.package("multiseq"),"extdata"))
samplesheet="samplesheetEncode.txt"
samplesheetToTrackHub(samplesheet, region="chr1:100-3000")
}  
}

