% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setAshParam}
\alias{setAshParam}
\title{Set default \code{ash} parameters.}
\usage{
setAshParam(ashparam)
}
\arguments{
\item{ashparam:}{a list of parameters to be passed to ash.}
}
\description{
Set default \code{ash} parameters.
}
\keyword{internal}

