% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{vss}
\alias{vss}
\title{Compute approximately unbiased variance estimates for the estimators for logit(p) when n is small}
\usage{
vss(n, s, f)
}
\arguments{
\item{n:}{number of trials}

\item{s:}{number of successes}

\item{f:}{number of failures}
}
\description{
Compute approximately unbiased variance estimates for the estimators for logit(p) when n is small
}
\keyword{internal}
