% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.bed}
\alias{write.bed}
\title{Write effect intervals to a \code{bed} file.}
\usage{
write.bed(intervals, bedfile)
}
\arguments{
\item{intervals:}{output of \code{\link{get.intervals}}}

\item{bedfile:}{output \code{bed} file}
}
\examples{
\dontrun{
data("dat",package="multiseq")
res <- multiseq(x=dat$x, g=dat$g, minobs=1, lm.approx=FALSE, read.depth=dat$read.depth)
intervals <- get.intervals(res, region=dat$region, what="effect")
write.bed(intervals, "out.bed")
}
}

