% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.gz}
\alias{write.gz}
\title{Write \code{\link{multiseq}} results to a compressed file.}
\usage{
write.gz(res, file = "results.mean.sd2.gz", what = "effect")
}
\arguments{
\item{res:}{\code{\link{multiseq}} output.}

\item{file:}{path to the output file.}

\item{what:}{if \code{what} is "log_baseline" then write \code{\link{multiseq}} \code{res$baseline.mean} and \code{res$baseline.var}; if \code{what} is "effect" then write \code{\link{multiseq}} \code{res$effect.mean} and \code{res$effect.var} output to a compressed file; defaults to "effect".}
}
\description{
The output file has two columns: first column the effect (or the baseline) mean and second column is the effect (or the baseline) posterior standard deviation ^2.
}
\examples{
\dontrun{
#run multiseq on sample data
data(dat, package="multiseq")
res <- multiseq(x=dat$x, g=dat$g, minobs=1, lm.approx=FALSE, read.depth=dat$read.depth)
write.gz(res, "./results.mean.sd2.gz")
}
}

