% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.summary}
\alias{write.summary}
\title{write.summary (to be removed)}
\usage{
write.summary(res, dir.name, timing, region)
}
\arguments{
\item{res:}{multiseq output}

\item{dir.name:}{output directory}

\item{timing:}{an integer indicating the running time}
}
\description{
File summary has 6 columns: chr, locus start, locus end, number of bases with effect at 2sd, number of bases with effect at 3sd, running time of multiseq
}

