% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{writeTrackHubSkeleton}
\alias{writeTrackHubSkeleton}
\title{Write genomes.txt and hub.txt files}
\usage{
writeTrackHubSkeleton(hub_dir, assembly = "hg19", hub_name_string,
  email = "esterpantaleo@gmail.com")
}
\description{
Write genomes.txt and hub.txt files
}

